[MOVE-ONLY] move CBudgetProposal/CFinalizedBudget to separate files
