+++ CMakeLists.txt
@@ -257,7 +257,9 @@ set(WALLET_SOURCES
         ./src/wallet/db.cpp
         ./src/addressbook.cpp
         ./src/crypter.cpp
+        ./src/budget/budgetproposal.cpp
         ./src/budget/budgetvote.cpp
+        ./src/budget/finalizedbudget.cpp
         ./src/budget/finalizedbudgetvote.cpp
         ./src/masternode.cpp
         ./src/masternode-budget.cpp
