+++ src/Makefile.am
@@ -205,7 +205,9 @@ BITCOIN_CORE_H = \
   logging.h \
   legacy/validation_zerocoin_legacy.h \
   sapling/sapling_validation.h \
+  budget/budgetproposal.h \
   budget/budgetvote.h \
+  budget/finalizedbudget.h \
   budget/finalizedbudgetvote.h \
   memusage.h \
   masternode.h \
@@ -376,7 +378,9 @@ libbitcoin_wallet_a_SOURCES = \
   addressbook.cpp \
   crypter.cpp \
   key_io.cpp \
+  budget/budgetproposal.cpp \
   budget/budgetvote.cpp \
+  budget/finalizedbudget.cpp \
   budget/finalizedbudgetvote.cpp \
   masternode.cpp \
   masternode-budget.cpp \
